Wall-Crawling
Player now uses raycasting to interact with the room's various surfaces, allowing them to change the incline of their movement (by either choosing to walk/crawl on the floor, or crawl on the walls or ceiling). The player then rotates to "stand" on that surface.
If trying to move from a "crawling" to a "standing" stance while on the walls or ceiling, the player will fall to a standing position on the floor.