Pickup Prefabs
Made a .blend file for small pickups (coins, gems) and saved to Unity project assets.
Created placeholder prefabs from those models.
In RoomBuilder script, created an empty with a "CoinParent" script to spawn and parent coins/gems to.
Also in RoomBuilder, created an empty with a "LaserParent" script to get that process started.