Pickup Tweak and Laser Setup
Pickups no longer spawn on top of each other.
Added laser node-, receiver-, and beam-spawning to LaserParent script. Despite much time and frustration, the lasers are implemented well enough to play with.
Added LaserBeam script to Beam prefabs to detect (trigger) collision with player, at which point the countdown timer placed on the LaserParent is activated. At zero seconds, the game is "over".