Doors
Modeled "Doorway" prefab and saved in "ModularPrefabs" assets.
Made "_Door" prefab and assigned to "RoomBuilder" script.
Door seemed too tall (because the player is so small), so shrunk wall and door models down to 2 units tall.
Added roomHeight variable to RoomBuilder, to make things easier if I decide to change the Y-axis scale of the room(s) in the future. Made roomHeight public, as it needs to be referenced by the LaserParent.