Laser Respawning
When the laser beams are spawning, the LaserParent script now determines if its collider is "touching" the Player (GameObject). If so, that Beam, along with its paired Node and Receiver.
For each "set" of laser modules which is destroyed in this way, a temporary value (respawnCount) is incremented and stored. If this value is greater than 0, the LaserParent script will run through the whole spawn cycle (SetSpawnPoints, SpawnNodes, SpawnReceivers, SpawnBeams, and check respawnCount again) over and over, until the room is fully populated by a number of lasers equal to the originally-assigned value (spawnCount).
Also, I spent a fuckload of time on this, getting it to work, but also cleaning up the script as a whole. Also, I added notes for possible tweaks to make in the future when I may be more capable, motivated or have more time to attempt than I do at the moment).