More "beamBlockers" and Moved Pickup Scripting
Through heavy trial-and-error with the Script Execution Order and shifting of script methods (Mostly between Awake, Start, and Update methods to get things to "line up", like making sure an array/list spawns BEFORE it starts being populated), I think I have managed to populate RoomFiller.beamBlockers well enough to work.
The list itself now contains the level's Player, doorways, and pickups (coins, gems), and considers them when determining if a laser spawns in a "conflicted" space.

During this process, I moved the scripting for Pickups from the PickupParent script to the RoomFiller script, as that will be a better place to hold the room's various objects (Player, pickups, display cases, furniture, and especially anything else to be added to "beamBlockers") in one place.