Door Prefab and Light Color Lerp
Modeled "_Door" and made a prefab.
When activated by the Player, the door swings open.
Created a quad on the other side of the door which takes the Player to the "GameOver" scene.

I tried for a fucking hour or more to get the lights in the scene to lerp their color back-and-forth from red to white. After fucking with it for WAY TO FUCKING LONG and getting basically nowhere, I scrapped the code. Still, the lights DO lerp to red when the Player touches a laser. Maybe that's enough for now.