Display Case Puzzle Management
Added "Puzzle Manager" script to DisplayCase prefab.
In this script, several states are monitored, including:
"In Progress", which disables player movement and the Laser Countdown Timer while the player attempts to solve the puzzle.
"Unsolved", which allows the player to abandon the puzzle (temporarily or permanently) and return to traversing the level. They are allowed to re-attempt the puzzle if they like.
"Solved", which rewards the player with points and destroys the individual Display Case's Puzzle Manager.
"Failed", which also destroys the Case's Puzzle Manager, but rewards the player with nothing.

When the player is done interacting with the puzzle, they are again allowed to move. The Laser Countdown Timer also resumes its previous state.

In the PlayerMove script, the player's forward raycast checks to see if a "reached" display case still has a Puzzle Manager script attached to it, so that the puzzle cannot be re-attempted if the player solves, or fails, the puzzle.