Glass Cutter 3
The Glass-cutting puzzle can now be left "unsolved" (not solved, but also not failed. Just abandoned). This does not destroy the puzzle script or any of the child-objects associated with it. It really just disables the puzzle's camera and re-enables the "PlayerCam".
Also did a bit of cleanup on the puzzle's script.