Alarm Box 2
Continued work on implementing Keypad and related scripting. It's basically done.
Buttons are spawned, positioned, and given strings of digits (one of which is the correct code to unlock the keypad). Then, upon being clicked, they display the placement of the correctly-placed digits in the sequence. If the correct code is chosen, the puzzle is solved.
Various outcomes occur, based on whether the puzzle is solved, failed, or left unsolved.
Cleaned up the various manager scripts, as they were still referencing the LaserParent OnEnable (moved to Awake).