Clamped Player Camera
Player camera is now clamped to a 180-degree rotation in front of the player, so that the player doesn't seem to be bending backward or doubling forward as the camera continues to rotate. There is some jittering at the extents, but I will fix that later.

Also did some cleanup on other scripts, to get them to the point that I could clear some space in my MonoDevelop panel. Having less scripts open might help to focus.