Implementing End-of-Level Achievements & Re-Implementation of "Locks" Scripts
Added a "WinBox" to the end of the level, to end the current heist and tally the player's achievements.

Added a "current time || par time" stopwatch text object to the player's UI, so they can see their progress at earning the relevant achievement.

Added "Money earned", "Achievement multiplier text" and "Achievements earned" texts to the level's "Win UI".

Added parameters for end-of-level achievements, which increase the "Achievement Multiplier", which in turn increases the player's score/earnings at the end of a heist.
So far, achievements are as follows:
-Disabled all alarms
-Collected all treasures/pickups
-No alarms activated
-Level completed below "par" time
-All Achievements Earned (This basically doubles the previous multipliers, allowing the player to earn the highest possible score/money in the current level).

Worked on spawning the right number of lasers and evening out the weight of their spawning (to prevent "clumping").

Tweaked "Door" prefab to spawn more easily in the correct position. Also refactored/re-implemented the previously-used "Lock" scripts used on the doors between rooms. Everything seems to work OK. Just needs some tweaks and visual improvements.